feat(in-person): pack picker + admin order guard (PR 7/7) (#48)
* feat(in-person): pack picker + admin order guard (PR 7/7)

- Add PackVariantPickerDialog: N slot pickers for pack products in
  in-person sales, repeat-aware stock, all-or-nothing confirm, splitPackPrice
  per slot
- Detect pack products in product search and route to pack picker
- Add Pack xN badge in product search dropdown for pack products
- Fetch pack_units from products table in in-person sales
- Export isPackProduct() helper from admin orders for order form guard
- Stock validation handled by existing BEFORE INSERT trigger on
  in_person_sale_items (no new validation needed)

* fix(storefront): pack pricing charges once at checkout

- fix(catalog): mapProduct now maps pack_units (product edit kept pack mode)
- fix(cart): pack-aware subtotal via computeCartSubtotal (no double charge)
- fix(checkout): OrderSummary shows split pack prices (created_at,id ordering)
- fix(rpc): remainder uses numeric division (::numeric / 100)
- fix(in-person): pack picker re-initializes slots on product change
- feat(catalog): totalStock aggregation + product card stock badge

+++ supabase/migrations/00004_bank_transfer_settings.sql
@@ -132,7 +132,7 @@ begin
            then floor(base * 100 / pack_units) / 100
            else base end as unit_price,
       case when pack_units is not null
-           then mod((base * 100)::int, pack_units) / 100
+           then mod((base * 100)::int, pack_units)::numeric / 100
            else 0 end as remainder
     from priced
   )
@@ -200,7 +200,7 @@ begin
            then floor(base * 100 / pack_units) / 100
            else base end as unit_price,
       case when pack_units is not null
-           then mod((base * 100)::int, pack_units) / 100
+           then mod((base * 100)::int, pack_units)::numeric / 100
            else 0 end as remainder
     from priced
   )
